# Optimize bot trading parameters for improved market making