--- a/.gitignore
+++ b/.gitignore
@@ -96,3 +96,5 @@ coverage.xml
 .env
 
 **/.claude/settings.local.json
+
+bots/