--- a/hummingbot/connector/exchange/xt/dummy.pxd
+++ b/hummingbot/connector/exchange/xt/dummy.pxd
@@ -1,0 +1,2 @@
+cdef class dummy():
+    pass