--- a/hummingbot/connector/exchange/xt/dummy.pyx
+++ b/hummingbot/connector/exchange/xt/dummy.pyx
@@ -1,0 +1,2 @@
+cdef class dummy():
+    pass