… README.md
… bots/deps-btc-dev/conf/conf_client.yml
… bots/deps-btc-dev/conf/strategies/deps-btc-conf_pure_mm.yml
… bots/deps-btc/conf/conf_client.yml
… bots/deps-btc/conf/strategies/deps-btc-conf_pure_mm.yml
… bots/deps-usdt-dev/conf/conf_client.yml
… bots/deps-usdt-dev/conf/strategies/deps-usdt-conf_pure_mm.yml
… bots/deps-usdt/conf/conf_client.yml
… bots/deps-usdt/conf/strategies/deps-usdt-conf_pure_mm.yml
… bots/deuro-btc-dev/conf/conf_client.yml
… bots/deuro-btc-dev/conf/strategies/deuro-btc-conf_pure_mm.yml
… bots/deuro-btc/conf/conf_client.yml
… bots/deuro-btc/conf/strategies/deuro-btc-conf_pure_mm.yml
… bots/deuro-usdt-dev/conf/conf_client.yml
… bots/deuro-usdt-dev/conf/strategies/deuro-usdt-conf_pure_mm.yml
… bots/deuro-usdt/conf/conf_client.yml
… bots/deuro-usdt/conf/strategies/deuro-usdt-conf_pure_mm.yml
… docker-compose.prod.yml
… docs/docker-multi-bot-quickstart.md
… hummingbot/VERSION
… hummingbot/connector/exchange/xt/xt_exchange.py
… hummingbot/connector/exchange/xt/xt_order_book.py
… hummingbot/connector/exchange/xt/xt_utils.py
… hummingbot/connector/exchange/xt/xt_web_utils.py
… hummingbot/core/rate_oracle/sources/coin_gecko_rate_source.py
… hummingbot/data_feed/coin_gecko_data_feed/coin_gecko_data_feed.py
… hummingbot/data_feed/custom_api_data_feed.py
… hummingbot/strategy/pure_market_making/pure_market_making.pxd
… hummingbot/strategy/pure_market_making/pure_market_making.pyx
… hummingbot/strategy/pure_market_making/pure_market_making_config_map.py