## Summary
This PR optimizes the trading parameters across all Hummingbot market making bots to improve responsiveness and trading efficiency.

## Changes Made

### 1. Reduced Order Refresh Times (All Bots)
- **order_refresh_time**: 60s → 30s
- Bots now check and update orders every 30 seconds instead of 60
- Faster adaptation to market price changes

### 2. Reduced Fill Order Delay (All Bots)
- **filled_order_delay**: 60s → 5s  
- After an order is filled, new orders are placed after only 5 seconds
- Increases trading volume and captures more opportunities

### 3. Standardized Spreads by Asset Type
- **USDT pairs** (DEURO-USDT, DEPS-USDT): 0.05% spread
- **BTC pairs** (DEURO-BTC, DEPS-BTC): 0.2% spread (remains unchanged)
- Tighter spreads for stable pairs, wider for volatile BTC pairs

### 4. Unified Refresh Tolerance
- **order_refresh_tolerance_pct**: All bots now use 0.05%
- More sensitive to price changes, triggering order updates sooner

## Affected Bots
- ✅ DEURO-USDT (prod & dev)
- ✅ DEURO-BTC (prod & dev)
- ✅ DEPS-USDT (prod & dev)
- ✅ DEPS-BTC (prod & dev)

## Testing
These parameters have been reviewed and optimized based on:
- EUR/USD stability for DEURO pairs
- DEPS token volatility characteristics
- BTC market dynamics
- Exchange rate limits and API constraints

## Impact
- More competitive spreads for USDT pairs
- Faster market response times
- Increased trading volume potential
- Better liquidity provision